Added Features and Reorganized Code
Added Keyboard Click enable function
Keyboard resets to initial values on reset
Reorganized code to make it easier to fill in the keymap. You can use a
spreadsheet editor with the template case X: addToKeyStack( Y );
break;
replacing X with the Sun keyboard hex code for the key and Y with the
Teensy library macro for the key.